requirements.txt
Flask==2.0.1
Flask_Cors==3.0.9
pymongo==3.11.0
python-dotenv==0.20.0
